pyproject.toml
[tool.poetry]
name = "drf-kit"
version = "1.0.0"
description = "DRF Toolkit"
authors = ["eduK <pd@eduk.com>"]
packages = [
    { include = "drf_kit" },
]


[tool.poetry.dependencies]
python = "^3.8"
Django = ">3,<3.1"  # while drf-extensions still remain incompatible
djangorestframework = "^3"
drf-extensions = "^0"
django-filter = "^2"
django-ordered-model = "^3"


[tool.poetry.dev-dependencies]
pylint = "*"
nose = "^1"
coverage = "*"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
